+++ src/helpers/index.js
@@ -15,5 +15,4 @@ export {
 	getReducers,
 	getMiddleware,
 	getAppWrappers,
-	createMasterEnhancer,
 } from './consolidators';
